+++ src/page-modules/contact/index.ts
@@ -4,4 +4,4 @@ export { type Line } from './server/journey-planner/validators';
 export { shouldShowContactPage } from './utils';
 export { MeansOfTransportContent } from './means-of-transport';
 export { default as TicketControlPageContent } from './ticket-control';
-export { default as TicketsAppContent } from './tickets-app';
+export { default as TicketingContent } from './ticketing';
