+++ src/page-modules/contact/ticketing/forms/priceAndTicketTypesForm.tsx
@@ -1,7 +1,7 @@
 import { SectionCard } from '../../components/section-card';
 import { PageText, useTranslation } from '@atb/translations';
 import { ticketsAppFormEvents } from '../events';
-import { TicketsAppContextProps } from '../tickets-app-form-machine';
+import { TicketsAppContextProps } from '../ticketingStateMachine';
 import { Typo } from '@atb/components/typography';
 import { Textarea } from '../../components/input/textarea';
 import { FileInput } from '../../components/input/file';
